Datastore client location refactor
